osclib: ignore_command: be more explicit that the comment is coming from an automated process
Adding a comment with the reason the request is being ignored is convenient, but
might be prone to misunderstandings especially since the reason is straight up
posted as a comment as well.

With this change, the comment is now posted using a template that makes
more explicit that it's an automated comment while keeping the actual reason
in the staging dashboard untouched.

Signed-off-by: Eugenio Paolantonio <eugenio.paolantonio@suse.com>